Prevent assignment to tuple and fix default expression
Creating a tuple as a target for a named expression was causing invalid
ASTs. There are probably other expressions that will cause this issue as
well.

The default expression was an ast.Name with id "None", and apparently
this is not valid. For "None", an ast.Constant should be used
instead. Python3.8 just started complaining about that after a recent
update. Now, the default expression is a name with "n" as the id.