General bug fixes
Test if the parent of the starred expression is an assignement before
checking if it's an only child. If it's not an assignment then we are
automatically fine because there must be another child somewhere (I
think).

Make the annotated assignment parent of an attribute non simple. (There
may be others assignable nodes that are valid in an annotatiion that
need this too.

Stop passing a nonsense argument to recursively_fix_context. It
wasn't used in this case and that's why it slipped by previously, but
now it's fixed.

Validate await is within an async function (As opposed to any function)