Create a new renderer
Making it complete and fully functional would probably take quite a lot
of work, so it's currently just an idea really. The main benefit is that
it would be made aware of the screen width and adapt the rendering to
fit in it (why leave it fixed at 80 columns when we have the power to
adapt?). Would also hopefully address some of the readability problems
with astor.

This adds an extra screen width parameter to the renderers which is
ignored by all of them except this new one (strender).

Also moves the initalization logic of main.py to the end so the
functions are already defined when in runs.