Make alternative function call command
This one calls a function on the selected node, whereas the original one
called the selected node. E.g. if you have "x" selected, the old one
would make x(n) while the new one makes f(x). I chose cC as the binding
to make it indicate that it's an alternative to cc.